Refactor component file structure for deprecation
Renamed several component files with an 'old' prefix, indicating a transition phase or deprecated status. Updated imports to reflect these changes. This preparatory step helps delineate old components from current or future implementations, facilitating code maintenance and future refactoring efforts.
