+++ src/components/index.ts
@@ -1,11 +1,11 @@
 export * from './ExternalLink';
-export * from './header';
-export * from './imageHeader';
-export * from './layout';
+export * from './oldheader';
+export * from './oldimageHeader';
+export * from './oldlayout';
 export * from './ListTags';
 export * from './PassWords';
 export * from './CodePen';
-export * from './posts';
-export * from './seo';
-export * from './sitemap';
+export * from './oldposts';
+export * from './oldseo';
+export * from './oldsitemap';
 export * from './Tags';
